+++ src/main/java/com/transport/management/modules/conductor/ConductorRepository.java
@@ -1,8 +1,7 @@
-package com.transport.management.repositories;
+package com.transport.management.modules.conductor;
 
 import java.util.List;
 
-import com.transport.management.modules.conductor.ConductorEntity;
 import com.transport.management.utils.abtractBase.BaseRepository;
 
 public interface ConductorRepository extends BaseRepository<ConductorEntity> {
