+++ src/main/java/com/transport/management/services/ConductorService.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.transport.management.modules.conductor.ConductorEntity;
-import com.transport.management.repositories.ConductorRepository;
+import com.transport.management.modules.conductor.ConductorRepository;
 import com.transport.management.utils.abtractBase.BaseService;
 
 @Service
